Refactor IQ to use StructOpt
Remove explicit clap dependency.
Use cargo environment
variables (https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-crates)
to get bin name and crate version information. This requires building with
cargo, but that's probably not a concern. If it is, we can use the version that
returns an Option<String> and have an explicit default fallback.