Introduce StructOpt and refactor cargo pants (#51)
* Introduce StructOpt and refactor cargo pants

Replicate the existing CLI experience from clap, but without the stringly-typed
interface.

* Refactor IQ to use StructOpt

Remove explicit clap dependency.
Use cargo environment
variables (https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-crates)
to get bin name and crate version information. This requires building with
cargo, but that's probably not a concern. If it is, we can use the version that
returns an Option<String> and have an explicit default fallback.

* Extract common log level code

* license headers

* Apply suggestions from code review

Co-authored-by: Dan Rollo <drollo@sonatype.com>

* Update src/bin/iq/main.rs

Co-authored-by: Dan Rollo <drollo@sonatype.com>

Co-authored-by: Jeffry Hesse <5544326+DarthHater@users.noreply.github.com>
Co-authored-by: Dan Rollo <drollo@sonatype.com>