Replace log4rs with tracing
log4rs has a transitive dependency on traitobject that has a vulnerability.
There hasn't been a release of log4rs in 11 months, so the vulnerability will
not likely be addressed in the near future.

tracing is a logging/tracing framework from the Tokio project that is compatible
with the output of the log crate, but provides additional constructs such as
spans to make debugging easier compared to correlating lines of output.

This change should be approximately equivalent to the previous solution. An
added benefit is that five verbose arguments ("-vvvvv") cause us to read the
RUST_LOG environment variable for a tracing directive (documented:
https://docs.rs/tracing-subscriber/0.3.1/tracing_subscriber/filter/struct.EnvFilter.html#directives).
This directive can be used to target a specific subsection of the program or a
dependency, without turning up the logging level for everything (e.g.
"RUST_LOG=debug,reqwest=trace" to investigate a network issue).

Additionally, this change includes some code cleanup changes such as removing
the old crate macro import syntax for serde and removing a ternary macro.

Signed-off-by: Amelia Keibler <akeibler@sonatype.com>